Remove single card builds
In order to keep the grouping logic as simple as possible we can't allow
for single card builds. A single card build acts as a way to get around
the single root card allowed in a group. If you wanted to group two
singles you would simply convert one of them to a single card build.

Since there isn't any known use case for single card builds, it is
probably best to explicitly prevent them for now.