Add pile ownership flag
This will allow us to keep track of the last person to add to a pile. If
we use this flag in conjunction with the length of a pile, then we will
be able to validate that each player only owns a single pile on the
floor at any given time.