Update InvalidHandAddressPosition rule
We can allow a hand address to be the move destination so long as we do
an implicit discard anytime that happens. This allows for the hand
address to appear as the first address in an annotation. That being
said, the hand address still needs to be last for pair moves.

Implementing this requires an implicit `collapse_floor` call inside of
the `apply_move` function. This is a breaking API change, but it is
needed for more than just the new hand address flexibility. In order to
properly implement a game of Suipi, you must collapse the floor after
every single move is applied. Allowing developers to skip this step can
result in a broken game.