Fix 4 + Ace + 3 is an invalid pile bug
When running an action to build an even numbered value, you can run into
an issue where two piles of equal value are passed to Pile::build.

This resulted in the PileError::BuildEqualValues check to pass,
preventing the action from being completed.

As an example, if you applied an annotation that represented building a
4 with and Ace and a 3, the 3 and Ace would combine to make a 4. The two
4s would then fail to be combined to make an 8.

To resolve this issue, we will bypass the equal values check unless both
piles are single card piles.