Fix `collapse_floor` stacking destination bug
When stacking two floor piles, the `collapse_floor` call towards the end
of the `apply` method can shift the destination address's value before
`validate_turn` is called.

This bug probably happens all the time. But it only causes issues if the
new value of the destination address is not available in your hand. In
that specific case we get a `StateError::UnpairablePileValue` result.
Marking the turn as invalid.

This bug was introduced by e10b38e, which added the ability to use a
hand address as the destination of a move. We handle this behavior by
doing an implicit discard at the end of the move. In order to properly
handle that `discard` call, we must first call `collapse_floor`.

To mitigate this issue we can move the main `collapse_floor` call to run
after `validate_turn`. This allows the `validate_turn` call to run with
an un-collapsed floor state. Then we will need to add a secondary
`collapse_floor` call immediately before the implicit discard of hand
address destinations. When the destination is a hand address, the check
for `StateError::UnpairablePileValue` is essentially a no-op. There is
never a case where a hand address's value is not included in the
player's hand. So in that case it is okay to call `collapse_floor`
before we run `validate_turn`.