Fix api get_scores during the middle of a game
Because the `Status.game` is 0 during the first game, the loop to get
the current scores would never run. This results in mid-game scores
being 0s until the end of the first game.

We can fix this by iterating over the `Game.scores` instead of looping
based on the current `Status.game` value.