Fix 2 + 3 + 2 is a valid build rule
There is an inconsistency in our rule based on the order you use two
cards with the same value. Something like 3 + 2 + 2 is invalid, because
you can't build 2 + 2 to get 4. But if you swap around the order that
the cards are built, you can have a valid move of 2 + 3 + 2.

In order to properly enforce the PileError::BuildEqualValues rule, we
should loop through all cards in all piles and make sure each of their
values are unique.

This implementation is also a proper fix for the "4 + Ace + 3 is an
invalid pile bug" from: 4ece3b7dcdf256e5a324d5b8c880e8ffa077c1bc.