Add cannot orphan owned pile rule
When you build a pile, or add to an existing one, you become the owner
of that pile. So that piles don't become orphaned, someone needs to be
able to pair each pile at all times. Pile ownership allows us to achieve
this goal by assigning a responsibility to be capable of pairing a pile
to a single player.

This, first and foremost, prevents a player from creating piles that
they cannot pair. But in the event that they do create a pile, that they
can pair, they should not be able to alter their hand such that they
cannot pair it in the future.

This rule should be enforced in the `State::validate_turn` method; such
that, a game state is never created with an orphaned pile.