Add point cards are not dealt to the floor rule
None of the following cards should be dealt to the floor as one of the
initial 4 cards of a game: 10♦, 2♠, A♠, A♣, A♥, A♦.

Each of these cards are worth points, and should not be up for grabs on
the first turn of the game. If one of these cards are dealt to the
floor, they should be placed at the bottom of the deck. This is the same
way we recycle cards with duplicate values that are dealt to the floor.